Move FileBlueprint tests into Blueprints/FileBlueprintTest
